Extract count measurements to separate classes
